Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a;
-    scanf("%d",a);
+    float a;
+    scanf("%f",a);
     float ar;
-    int b = 3.14;
+    float b = 3.14;
     ar = b*a*a;
     printf("Area:%.2f",ar);
     return 0;
